Move github_login.py to integrations/github/login.py
