+++ tests/integrations/test_github_login.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 import pytest
 
-from integrations import github_login
+from integrations.github import login as github_login
+from integrations.github.login import GitHubLoginResult
 from integrations.github_identity import saved_github_username
-from integrations.github_login import GitHubLoginResult
 from integrations.github_mcp import GitHubMCPValidationResult
 from integrations.github_mcp_oauth import GitHubDeviceToken
