+++ tests/interactive_shell/runtime/test_first_launch_github.py
@@ -5,8 +5,8 @@ import io
 import pytest
 from rich.console import Console
 
-from integrations import github_login as github_login_mod
-from integrations.github_login import GitHubLoginResult
+from integrations.github import login as github_login_mod
+from integrations.github.login import GitHubLoginResult
 from integrations.github_mcp import DEFAULT_GITHUB_MCP_TOOLSETS, DEFAULT_GITHUB_MCP_URL
 from integrations.github_mcp_oauth import GitHubDeviceCode
 from platform.analytics import source as analytics_source
